+++ Sources/NablaCore/Modules/VideoCallModule.swift
@@ -1,14 +1,15 @@
 import Foundation
+import UIKit
 
 public protocol VideoCallModule: Module {
     func makeClient(container: CoreContainer) -> VideoCallClient
 }
 
 public protocol VideoCallClient {
-    var views: VideoCallViewFactory { get }
+    var crossModuleViews: VideoCallViewFactory { get }
     
     // Ideally we would extract VideoCall cells to NablaVideoCall package to avoid such public APIs, see #20836 for more details
     var currentVideoCallToken: String? { get }
     func watchCurrentVideoCall(callback: @escaping (_ token: String?) -> Void) -> Cancellable
-    func openCurrentVideoCall()
+    func openVideoCallRoom(url: String, token: String, from viewController: UIViewController)
 }