+++ Sources/NablaCore/Modules/VideoCallViewFactory.swift
@@ -1,5 +1,3 @@
 import UIKit
 
-public protocol VideoCallViewFactory {
-    func createVideoCallRoomViewController(url: String, token: String) -> UIViewController
-}
+public protocol VideoCallViewFactory {}