+++ Sources/NablaVideoCall/NablaVideoCallClient+Views.swift
@@ -1,7 +1,11 @@
 import NablaCore
 
 public extension NablaVideoCallClient {
-    var views: VideoCallViewFactory {
-        NablaVideoCallViewFactory(client: self)
+    var views: NablaVideoCallViewFactory {
+        NablaVideoCallViewFactoryImpl(client: self)
+    }
+    
+    var crossModuleViews: VideoCallViewFactory {
+        NablaVideoCallViewFactoryImpl(client: self)
     }
 }